temporarily suppress messages and enable multiple seasons per season (sub-seasons)...

+++ season.R
@@ -5,7 +5,7 @@ library(argparse)
 parser <- ArgumentParser(description = "Process broomball stats into a season report...")
 parser$add_argument("db", help = "the path to the broomball database")
 parser$add_argument("year", help = "the path to the assay result directory", type = "integer")
-parser$add_argument("season", help = "the path to the assay result directory", type = "integer")
+parser$add_argument("season", help = "the path to the assay result directory", type = "double")
 parser$add_argument("-out", help = "the output file name")
 args <- parser$parse_args()
